🛠 refactor: reports
