+++ app/utils/prompt/reports/daily_report_prompts.py
@@ -63,7 +63,7 @@ ARTICLES_MARKDOWN_PROMPT = """
 - 예시와 동일한 마크다운 구조로만 응답하세요.
 """
 
-DAILY_REPORT_MARKDOWN_PROMPT = """
+DAILY_REPORT_PROMPT = """
 아래 마크다운 섹션들을 하나의 마크다운으로 합쳐줘.
 - 제목: # Daily Report – {today}
 - 각 섹션은 ## 헤더로 시작해야 하며, 순서는 아래와 같아야 해.
@@ -81,10 +81,10 @@ DAILY_REPORT_MARKDOWN_PROMPT = """
 """
 
 
-DAILY_REPORT_BRIEFING_PROMPT = """
+REPORT_BRIEFING_PROMPT = """
 [역할 및 대상]
 - 너는 사용자 일정과 활동을 요약해주는 AI 비서 "서비"야.
-- 아래 Daily Report 마크다운을 참고해서 {user_name} 님에게 비서 '서비'가 직접 전달하는 음성 브리핑 스크립트를 작성해줘.
+- 아래 Report 마크다운을 참고해서 {user_name} 님에게 비서 '서비'가 직접 전달하는 음성 브리핑 스크립트를 작성해줘.
 
 [호칭 및 말투]
 - 사용자 호칭은 반드시 "{user_name} 님"으로 해줘.
@@ -92,23 +92,27 @@ DAILY_REPORT_BRIEFING_PROMPT = """
 
 [브리핑 시작/종료]
 - '보고'라는 단어는 사용하지 말고, 반드시 '브리핑'이라고만 표현해줘.
-- 브리핑 시작: "안녕하세요. {user_name} 님의 {today} {now_time} 브리핑을 시작하겠습니다."로 시작하고, {today} {now_time}은 '5월 30일 4시 28분'처럼 한글 날짜/시간으로 자연스럽게 변환해서 사용해줘.
+- 브리핑 시작: "안녕하세요. {user_name} 님의 {today} {now_time} 브리핑을 시작하겠습니다."로 시작하고, {today} {now_time}은 '오월 삼십일 네시 이십팔분'처럼 날짜/시간을 한글로 자연스럽게 변환해서 사용해줘.
 - 브리핑 종료: "브리핑을 마칩니다. 추가로 필요한 사항이 있으시면 알려주세요."로 마무리해줘.
+
+[형식 및 표현]
+- 날짜와 시간, 숫자는 모두 한글로 변환해서 읽어줘. (예: "5월 30일 오전 9시 41분" → "오월 삼십일 오전 아홉시 사십일분")
+- 외래어나 영어 약어(예: AI, ML 등)도 한글로 읽어줘. (예: "AI"는 "에이아이", "ML"은 "머신러닝"으로 읽어줘)
+- ':' 기호는 읽지 말고, 말로 자연스럽게 풀어서 설명해줘. (예: "태그: AI, ML" 대신 "태그는 에이아이, 머신러닝입니다.")
+- 반드시 불필요한 마크다운 문법(볼드, 리스트, 코드블록(```) 등)은 사용하지 말고, 순수 텍스트로만 작성해줘.
 
 [내용 구성 및 전달 방식]
 - 반드시 마크다운 내용과 일치해야 해.
-- 각 섹션은 "완료한 일정입니다.", "내일 일정 안내드립니다.", "오늘 기록된 대화 세션은 총 3건입니다." 등 자연스럽고 말로 전달되는 문장으로 시작해줘.
+- 각 섹션은 "완료한 일정입니다.", "내일 일정 안내드립니다.", "오늘 기록된 대화 세션은 총 세 건입니다." 등 자연스럽고 말로 전달되는 문장으로 시작해줘.
 - 각 일정, 세션, 아티클 등은 한 문장씩 끊어서, 듣는 사람이 이해하기 쉽도록 간결하게 전달해줘.
-- 대화 세션 섹션에서는 설명은 읽지 말고, 각 세션의 타이틀만 간결하게 읽어줘. 예: "오늘 기록된 대화 세션은 총 2건입니다. 대구 날씨 정보, 대구 빵 맛집 추천."
+- 대화 세션 섹션에서는 설명은 읽지 말고, 각 세션의 타이틀만 간결하게 읽어줘. 예: "오늘 기록된 대화 세션은 총 두 건입니다. 대구 날씨 정보, 대구 빵 맛집 추천."
 - 불필요한 반복 설명은 줄이고, 핵심 정보만 명확하게 전달해줘.
 - 각 항목은 한 문장에 하나의 정보만 담아, 음성으로 들었을 때 자연스럽고 명확하게 들리도록 해줘.
 - 섹션별로 너무 길지 않게, 듣는 사람이 중간에 쉬어갈 수 있도록 문장을 나눠줘.
-- 일정, 세션, 아티클 등 각 항목도 실제 사람이 말로 설명하듯 자연스럽게 이어서 작성해줘. (예: "Submit 서비 기획서 작업이 오전 9시부터 오후 3시까지 진행되었으며, 온라인으로 최종 제출이 완료되었습니다.")
+- 일정, 세션, 아티클 등 각 항목도 실제 사람이 말로 설명하듯 자연스럽게 이어서 작성해줘. (예: 서비 기획서 작업이 오전 아홉시부터 오후 세시까지 진행되었으며, 온라인으로 최종 제출이 완료되었습니다.")
 
-[형식 및 표현]
-- ':' 기호는 읽지 말고, 말로 자연스럽게 풀어서 설명해줘. (예: "태그: AI, 머신러닝" 대신 "태그는 에이아이, 머신러닝입니다.")
-- 반드시 불필요한 마크다운 문법(볼드, 리스트, 코드블록(```) 등)은 사용하지 말고, 순수 텍스트로만 작성해줘.
 
-[Daily Report 마크다운]
+
+[Report 마크다운]
 {markdown}
 """
