+++ pyproject.toml
@@ -28,4 +28,5 @@ dependencies = [
     "psycopg2-binary>=2.9.10",
     "python-dotenv>=1.1.0",
     "tavily-python>=0.7.2",
+    "tzdata>=2025.2",
 ]
